updated delete folders AND associated notes

+++ routes/folders.js
@@ -110,10 +110,19 @@ router.put('/folders/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/folders/:id', (req, res, next) => {
   const { id } = req.params;
+ 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    const err = new Error('The `id` is not valid');
+    err.status = 400;
+    return next(err);
+  }
 
   Folder.findByIdAndRemove(id)
     .then(() => {
-      res.status(204).end();
+      Note.deleteMany( {folderId: id} )
+        .then(() => {
+          res.status(204).end();
+        });
     })
     .catch(err => {
       next(err);
